requirements.txt
Pillow==9.4.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
